+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 0.2.0 - 2022-08-26
+### Added
+    - added a new macro "BufferData" (see docs for more details)
+### Fix
+    - Fixed array_stride data to take only the size of provided attributes
+
 ## 0.1.2 - 2022-08-20
 ### Added
-    - added a changelod.md
+    - added a changelog.md
 ### Fix
     - Fixed errornous shader location
 ## 0.1.1 - 2022-08-18